+++ src/index.mjs
@@ -18,7 +18,8 @@ export { default as flatMap } from './flat-map'
 export { default as groupBy } from './group-by'
 export { default as iterable } from './iterable'
 export { default as keys } from './keys'
-export { default as largest } from './largest'
+export { default as nlargest } from './nlargest'
+export { default as nsmallest } from './nlargest'
 export { default as map } from './map'
 export { default as permutations } from './permutations'
 export { default as product } from './product'
